V2 from float to int in tf.cond

+++ baselines/pois/pois.py
@@ -530,7 +530,7 @@ def learn(make_env, make_policy, *,
     if entropy != 'none':
         scheme, v1, v2 = entropy.split(':')
         if scheme == 'step':
-            entcoeff = tf.cond(tf.less_equal(iter_progress_ * max_iters, float(v2)), lambda: float(v1), lambda: float(0.0))
+            entcoeff = tf.cond(tf.less_equal(iter_progress_ * max_iters, int(v2)), lambda: float(v1), lambda: float(0.0))
             losses_with_name.append((entcoeff, 'EntropyCoefficient'))
             entbonus = entcoeff * meanent
             bound_ = bound_ + entbonus
